Stopped deleting the ORCA dir if it's the same as the morphology one

+++ morphct/run_MorphCT.py
@@ -241,8 +241,11 @@ class simulation:
 
     def make_dir_tree(self):
         print("Sorting out directory structure...")
-        # Delete the ORCA information if the user has asked to
-        if self.overwrite_current_data is True:
+        # Delete the ORCA information if the user has asked to (as long as it's not in
+        # the same place as the morphology data!!)
+        if (self.overwrite_current_data is True) and (
+            self.output_orca_directory != self.output_morphology_directory
+        ):
             print("OVERWRITE CURRENT DATA IS TRUE...EMPTYING ORCA DIR")
             try:
                 shutil.rmtree(self.output_orca_directory)
